Removed repetition that JwtAuthenticationFilter does every request.

+++ src/main/java/com/mintmoose/springbootebay/Controller/CustomerController.java
@@ -3,12 +3,15 @@ package com.mintmoose.springbootebay.Controller;
 import com.mintmoose.springbootebay.Config.JwtService;
 import com.mintmoose.springbootebay.Model.Customer;
 import com.mintmoose.springbootebay.Model.NewCustomerRequest;
+import com.mintmoose.springbootebay.Model.Product;
 import com.mintmoose.springbootebay.Service.CustomerService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
@@ -30,40 +33,35 @@ public class CustomerController {
     }
 
     @DeleteMapping("/{customerUsername}")
-    public ResponseEntity<?> deleteCustomer(@PathVariable String customerUsername, @RequestHeader("Authorization") String authorizationHeader) {
-        return processCustomerRequest(customerUsername, authorizationHeader, (token, requestCustomer) -> {
-            customerService.deleteCustomer(requestCustomer.getCustomerId());
-            return ResponseEntity.status(HttpStatus.OK).body("Customer deleted successfully.");
-        });
-    }
-
-    @PutMapping("/{customerUsername}")
-    public ResponseEntity<?> updateCustomer(@PathVariable String customerUsername, @RequestBody NewCustomerRequest request, @RequestHeader("Authorization") String authorizationHeader) {
-        return processCustomerRequest(customerUsername, authorizationHeader, (token, requestCustomer) -> {
-            Customer updatedCustomer = customerService.updateCustomer(requestCustomer.getCustomerId(), request);
-            if (updatedCustomer != null) {
-                return ResponseEntity.ok(updatedCustomer);
-//                return ResponseEntity.status(HttpStatus.OK).body("Customer updated successfully.");
+    public ResponseEntity<?> deleteCustomer(@PathVariable String customerUsername, Authentication authentication) {
+        if (authentication.isAuthenticated()) {
+            String username = authentication.getName();
+            Customer requestCustomer = customerService.getCustomerByUsername(username)
+                    .orElseThrow(() -> new IllegalArgumentException("Access denied. Invalid authorization."));
+            try {
+                if (Objects.equals(customerUsername, username)) {
+                    customerService.deleteCustomer(requestCustomer.getCustomerId());
+                    return ResponseEntity.status(HttpStatus.OK).body("Customer deleted successfully.");
+                }
+            } catch (IllegalArgumentException e) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
             }
-            return ResponseEntity.notFound().build();
-
-        });
-    }
-
-    private ResponseEntity<?> processCustomerRequest(String customerUsername, String authorizationHeader, CustomerRequestProcessor processor) {
-        String token = authorizationHeader.substring(7);
-        Customer requestCustomer = customerService.getCustomerByUsername(customerUsername)
-                .orElseThrow(() -> new IllegalArgumentException("Customer not found."));
-
-        if (jwtService.isTokenValid(token, requestCustomer)) {
-            return processor.process(token, requestCustomer);
         }
-
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. Invalid authorization.");
     }
 
-    private interface CustomerRequestProcessor {
-        ResponseEntity<?> process(String token, Customer requestCustomer);
+    @PutMapping("/{customerUsername}")
+    public ResponseEntity<?> updateCustomer(@PathVariable String customerUsername, @RequestBody NewCustomerRequest request, Authentication authentication) {
+        if (authentication.isAuthenticated()) {
+            String username = authentication.getName();
+            Customer requestCustomer = customerService.getCustomerByUsername(username)
+                    .orElseThrow(() -> new IllegalArgumentException("Access denied. Invalid authorization."));
+            Customer updatedCustomer = customerService.updateCustomer(requestCustomer.getCustomerId(), request);
+            return ResponseEntity.ok(updatedCustomer);
+//          return ResponseEntity.status(HttpStatus.OK).body("Customer updated successfully.");
+        }
+        return ResponseEntity.notFound().build();
     }
+
 }
